+++ backend/users/views.py
@@ -25,8 +25,10 @@ class UserViewSet(viewsets.ModelViewSet):
             serializer = FollowSerializer(author,
                                           context={'request': request})
             return Response(serializer.data, status=status.HTTP_200_OK)
-        Follow.objects.filter(user=request.user, author=author).delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        elif request.method == 'DELETE':
+            Follow.objects.filter(user=request.user, author=author).delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return None
 
 
 class FollowViewSet(viewsets.ModelViewSet):
